Options to customize wine sandbox. Close #596

+++ lutris/util/wineprefix.py
@@ -1,8 +1,10 @@
 import os
+import time
 from lutris.util.wineregistry import WineRegistry
 from lutris.util.log import logger
 from lutris.util import joypad
 
+desktop_folders = ["Desktop", "My Documents", "My Music", "My Videos", "My Pictures"]
 
 class WinePrefixManager:
     """Class to allow modification of Wine prefixes without the use of Wine"""
@@ -12,7 +14,6 @@ class WinePrefixManager:
         self.path = path
 
     def setup_defaults(self):
-        self.sandbox()
         self.override_dll("winemenubuilder.exe", "")
 
     def get_registry_path(self, key):
@@ -48,14 +49,44 @@ class WinePrefixManager:
             return
         self.set_registry_key(key, dll, mode)
 
-    def sandbox(self):
+    def desktop_integration(self, desktop_dir=None):
+        """Overwrite desktop integration"""
+
         user = os.getenv('USER')
         user_dir = os.path.join(self.path, "drive_c/users/", user)
-        # Replace symlinks
+
+        if (not desktop_dir):
+            desktop_dir = user_dir
+
         if os.path.exists(user_dir):
-            for item in os.listdir(user_dir):
+            # Replace desktop integration symlinks
+            for item in desktop_folders:
                 path = os.path.join(user_dir, item)
-                if os.path.islink(path):
+                old_path = path + ".winecfg"
+
+                if (os.path.islink(path)):
+                    os.unlink(path)
+                elif (os.path.isdir(path)):
+                    try:
+                        os.rmdir(path)
+                    # We can't delete nonempty dir, so we rename as wine do.
+                    except OSError:
+                        os.rename(path, old_path)
+
+                if (desktop_dir != user_dir):
+                    src_path = os.path.join(desktop_dir, item)
+                    os.makedirs(src_path, exist_ok=True)
+                    os.symlink(src_path, path)
+                else:
+                    # We use first the renamed dir, otherwise we make it.
+                    if (os.path.isdir(old_path)):
+                        os.rename(old_path, path)
+                    else:
+                        os.makedirs(path, exist_ok=True)
+
+            # Security: Remove other symlinks.
+            for item in os.listdir(user_dir):
+                if not item in desktop_folders and os.path.islink(path):
                     os.unlink(path)
                     os.makedirs(path)
